requirements.txt
cn2an==0.5.22
Cython==0.29.21
g2p_en==2.1.0
gradio==3.45.1
jieba==0.42.1
librosa==0.9.1
matplotlib==3.7.1
numpy==1.23.5
pyopenjtalk==0.3.2
pypinyin==0.49.0
PySoundFile==0.9.0.post1
scipy==1.10.1
soundfile==0.12.1
tqdm==4.65.0
transformers==4.28.1
tensorboard==2.3.0
protobuf==3.19.0
openai-whisper